+++ R/hshrink.R
@@ -39,6 +39,9 @@
 #' @return The ranger object is modified in-place. 
 #'
 #' @examples
+#' ## Hierarchical shrinkage for a probablity forest
+#' rf <- ranger(Species ~ ., iris, node.stats = TRUE, probability = TRUE)
+#' hshrink(rf, lambda = 5)
 ##' @references
 ##' \itemize{
 ##'   \item Agarwal, A., Tan, Y.S., Ronen, O., Singh, C. & Yu, B. (2022). Hierarchical Shrinkage: Improving the accuracy and interpretability of tree-based models. Proceedings of the 39th International Conference on Machine Learning, PMLR 162:111-135.