+++ NEWS.md
@@ -1,4 +1,13 @@
 
+# ranger 0.17.0
+* New CRAN version
+
+# ranger 0.16.4
+* Add handling of missing values for classification and regression
+
+# ranger 0.16.3
+* Fix a bug for always.split.variables (for some settings)
+
 # ranger 0.16.2
 * Add Poisson splitting rule for regression trees